[AriochNYCSchoolsDemo] Adding a dependencies and network layer
Adding an initial network layer, it may evolve as the other layers are added however, for the time being this feels like a good start. It includes API endpoints for the 2018 nyc school data and the 2022 sat data.
-generic network handler that will take care of all the requests.
-network state interface that will provide loading, idle and not connected as the values for errors etc.
-network result which will wrap the network success and error responses. There is a network silent state which will be used as the initial state of the networkResult.
-Singleton retrofit instance.
-Instance of the NYCSchoolsApiService that will provide the data.
-Network specific data objects for the NYCSchools2018 data and the NYC Schools 2022 SAT Data. These objects will be converted into other objects for the higher layers.
There are a number of unused flows at the moment, they will get collected or turned into livedata objects for use in the UI.

Added dependencies for livedata, room, retrofit, lifecycle components and some others that are planned for this app.